move bt_clr because it might be causing me woes
I've had bluetooth stop working several times recently and I thiiiiink
it has something to do with a false trigger on layer tap with the
combination `SPC y` being registered as BT_CLR, so I'm going to move
BT_CLR to a key I don't otherwise use and see if that remedies it. This
is rather freaking annoying behavior of layer taps. I don't think I've
had quite so many problems with false layer taps on thumb keys on qmk,
so I dunno what the settings differences might be. will have to investigate