add a 3 finger combo for shifted accents
not entirely sure I'm going to love this but I also assume it will be
infrequent enough that it won't really matter. It *does* feel a little
more natural than the 2 finger combo plus then a hold for shift (with one
of the combo keys being held)

I do like that I can just hold the combo keys and it keeps the layer active
like it turns the combo into a hold. It's nice. Completely unnecessary as
I don't think there are any chained accent characters in Spanish, but whatev