move the LA to the parameter
not entirely sure why but the macro expansion didn't seem to like the LA()
surrounding the placeholder, so we'll just move it out to the parameter

I don't *love* this, but it does at least make this macro very generic for
any sort of deadkey type stuff where it's deadkey followed by regular key
though I suspect on macos that's mostly going to be LA(), it's fine. Totes fine.